Todo.Domain/Commands/Inputs/MarkTodoAsUndoneCommand.cs
using System;
using Flunt.Notifications;
using Flunt.Validations;
using Todo.Domain.Commands.Inputs.Contracts;

namespace Todo.Domain.Commands.Inputs
{
    public class MarkTodoAsUndoneCommand : Notifiable<Notification>, ICommand
    {
        public MarkTodoAsUndoneCommand() { }

        public MarkTodoAsUndoneCommand(Guid id, string user)
        {
            Id = id;
            User = user;
        }

        public Guid Id { get; set; }
        public string User { get; set; }

        public bool Validate()
        {
            AddNotifications(
                new Contract<Notification>()
                    .Requires()
                    .IsGreaterOrEqualsThan(User, 6, "User", "Usuario não pode conter menos de 6 caracteres!")
                    .IsNotNullOrEmpty(User, "User", "Usuario não pode ser nulo")
            );

            return IsValid;
        }
    }
}